chore: refactor lemminx extension structure
